# Vet Clinic App - Database Performance audit