--- a/data.sql
+++ b/data.sql
@@ -197,6 +197,8 @@ INSERT INTO animals (name, date_of_birth, escape_attempts, neutered, weight_kg) 
 INSERT INTO animals (name, date_of_birth, escape_attempts, neutered, weight_kg) VALUES ('Pikachu', '2021-01-07', '1', false, '15.04');
 
 INSERT INTO animals (name, date_of_birth, escape_attempts, neutered, weight_kg) VALUES ('Devimon', '2017-05-12', '5', true, '11.0');
+
+-- WEEK 2 - Database performance audit
 -- m
 -- This will add 3.594.280 visits considering you have 10 animals, 4 vets, and it will use around ~87.000 timestamps (~4min approx.)
 
@@ -205,5 +207,5 @@ INSERT INTO visits (animal_id, vet_id, date_of_visit) SELECT * FROM (SELECT id F
 
 -- This will add 2.500.000 owners with full_name = 'Owner <X>' and email = 'owner_<X>@email.com' (~2min approx.)
 
-insert into owners (full_name, email) select 'Owner ' || generate_series(1,2500000), 'owner_' || generate_series(1,2500000) || '@mail.com';
-insert into owners (full_name, email) select 'Owner ' || generate_series(2500001,3000000), 'owner_' || generate_series(2500001,3000000) || '@mail.com';
+INSERT INTO owners (full_name, email) SELECT 'Owner ' || generate_series(1,2500000), 'owner_' || generate_series(1,2500000) || '@mail.com';
+INSERT INTO owners (full_name, email) SELECT 'Owner ' || generate_series(2500001,3000000), 'owner_' || generate_series(2500001,3000000) || '@mail.com';