### [Description](https://github.com/microverseinc/curriculum-databases/blob/main/db-structure/vet_clinic_performance_audit.md#description)
In this project, you will have a chance to optimize some slow queries in your database.


### [General requirements](https://github.com/microverseinc/curriculum-databases/blob/main/db-structure/vet_clinic_performance_audit.md#general-requirements)
- Make sure that you used the correct [Gitflow](https://github.com/microverseinc/curriculum-transversal-skills/blob/main/git-github/articles/gitflow.md).
- Make sure that you documented your work [in a professional way](https://github.com/microverseinc/curriculum-transversal-skills/blob/main/documentation/articles/professional_repo_rules.md).

### [Preparation (0.5h)](https://github.com/microverseinc/curriculum-databases/blob/main/db-structure/vet_clinic_performance_audit.md#preparation-05h)
- For this project, you need special preparation. As the goal of this project is to solve some performance issues, first we need to introduce those issues. In order to do that, you will populate your database with a significant number of data.

- Please complete the following steps:

- Make sure that you have your database set up with the schema and data from your previous projects.
- Run the following query to add an extra column to the owners table:
-- Add an email column to your owners table
ALTER TABLE owners ADD COLUMN email VARCHAR(120);
- Run the following statements to add data to your database (executing them might take a few minutes):
-- This will add 3.594.280 visits considering you have 10 animals, 4 vets, and it will use around ~87.000 timestamps (~4min approx.)
INSERT INTO visits (animal_id, vet_id, date_of_visit) SELECT * FROM (SELECT id FROM animals) animal_ids, (SELECT id FROM vets) vets_ids, generate_series('1980-01-01'::timestamp, '2021-01-01', '4 hours') visit_timestamp;

-- This will add 2.500.000 owners with full_name = 'Owner <X>' and email = 'owner_<X>@email.com' (~2min approx.)
insert into owners (full_name, email) select 'Owner ' || generate_series(1,2500000), 'owner_' || generate_series(1,2500000) || '@mail.com';
- 4 Depending on your machine speed, it might be enough or not. Check that by running explain analyze SELECT COUNT(*) FROM visits where animal_id = 4: - If you get Execution time: X ms and X >= 1000: that should be enough, you can continue to the project requirements. - If you get Execution time: X ms and X < 1000: please go back to point 3. and repeat until you get a value bigger than 1000ms.

### [Project requirements (1h)](https://github.com/microverseinc/curriculum-databases/blob/main/db-structure/vet_clinic_performance_audit.md#project-requirements-1h)
- The following queries are taking too much time (1 sec = 1000ms can be considered as too much time for database query). Try them on your machine to confirm it:

SELECT COUNT(*) FROM visits where animal_id = 4;
SELECT * FROM visits where vet_id = 2;
SELECT * FROM owners where email = 'owner_18327@mail.com';
- Use EXPLAIN ANALYZE on the previous queries to check what is happening. Take screenshots of them - they will be necessary later.

- Find a way to decrease the execution time of the first query. Look for hints in the previous lessons.

- Find a way to improve the execution time of the other two queries.

- While you are making changes, check if they help by running EXPLAIN ANALYZE. Once you succeed, take a screenshot of the EXPLAIN ANALYZE result showing that time actually decreased.

- Any changes you made to your database schema should be added to the schema.sql file and committed in a new pull request.

- If you decide to commit any INSERT INTO queries - remember to add the m to the data.sql file. It is important to keep only the queries that change the database structure in the schema.sql file.

- In your pull request description include screenshots with EXPLAIN ANALYZE results (before and after you improve the speed) for each of the 3 problematic queries.

** QUERIES ANALYSIS AND SCREENSHOTS

### FIRST QUERY
The PostgreSQL dbms took  03.732 seconds to evaluate
EXPLAIN ANALYZE SELECT COUNT(*) FROM visits where animal_id = 4;
Since it took more than 1 second we tried to improve its performance creating an INDEX for the animal_id for the VISITS table.
The time after creating the index dropped to 00.346 seconds.
  ### BEFORE
![EXPL-query1](https://github.com/jcmunav63/vet_clinic_db_performance/assets/11914857/5d9e61bb-da37-49cd-8b2e-cd2322797434)
  ### AFTER
![EXPL-IMPR-query1](https://github.com/jcmunav63/vet_clinic_db_performance/assets/11914857/fd107d1b-4668-4cec-8626-66ba7afb78bd)

### SECOND QUERY
The PostgreSQL dbms took 04.501 seconds to evaluate
EXPLAIN ANALYZE SELECT * FROM visits where vet_id = 2;
Since it took more than 1 second we tried to improve its performance creating an INDEX for the vet_id for the VISITS table.
The time after creating the index dropped to 01.072 seconds.
   ### BEFORE
![EXPL-query2](https://github.com/jcmunav63/vet_clinic_db_performance/assets/11914857/640d459c-540a-438c-b703-d45e06cee8ff)
   ### AFTER
![EXPL-IMPR-query2](https://github.com/jcmunav63/vet_clinic_db_performance/assets/11914857/a53f0250-6886-4842-8c4c-6abd6be1494f)

### THIRD QUERY
We also created an INDEX for email in the OWNERS table. The query took more than 1 second at the beginning (02.363 seconds), but after creating the index it dropped to 00.258 seconds.
EXPLAIN ANALYZE SELECT * FROM owners where email = 'owner_18327@mail.com';
   ### BEFORE
![EXPL-query3](https://github.com/jcmunav63/vet_clinic_db_performance/assets/11914857/1abbc020-7a68-4d0d-9bb5-b376e6aacb70)
   ### AFTER
![EXPL-IMPR-query3](https://github.com/jcmunav63/vet_clinic_db_performance/assets/11914857/d01f8fea-4c4a-4c5f-be8c-86a28d201918)